Use int from futurize instead of long

+++ lib/pytis/dbdefs/db_pytis_menu.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8
 
 from __future__ import unicode_literals
+from builtins import range, int
 
-from builtins import range
 import sqlalchemy
 from pytis.data.dbdefs import and_, or_, not_, sval, func
 import pytis.data.gensqlalchemy as sql
@@ -540,7 +540,7 @@ class EPytisMenuTrigger(Base_PyTriggerFunction):
                                 next_suffix = next_item_position[-1]
                                 suffix += '0' * max(len(next_suffix) - len(suffix), 0)
                                 next_suffix += '0' * max(len(suffix) - len(next_suffix), 0)
-                                new_suffix = str(long((long(suffix) + long(next_suffix)) // 2))
+                                new_suffix = str((int(suffix) + int(next_suffix)) // 2)
                                 while len(new_suffix) < len(next_suffix):
                                     new_suffix = '0' + new_suffix
                                 if new_suffix == suffix:
@@ -555,7 +555,7 @@ class EPytisMenuTrigger(Base_PyTriggerFunction):
                             if suffix[-1] == '9':
                                 next_position = position[:-1] + [position[-1] + '4']
                             else:
-                                next_position = position[:-1] + [str(long(position[-1]) + 1)]
+                                next_position = position[:-1] + [str(int(position[-1]) + 1)]
                             update_next_position(position, next_position)
                 plpy.execute("select e_pytis_menu_check_positions()")
 
@@ -782,7 +782,7 @@ class PytisFirstPosition(Base_PyFunction):
     def pytis_first_position(position):
         position = args[0]
         start = '1' + '0' * (len(position) - 1)
-        first = str(long((long(start) + long(position)) // 2))
+        first = str((int(start) + int(position)) // 2)
         if first == start:
             first += '8'
         return first
